### 2024-05-30 #6806 @editorialbot
[PRE REVIEW]: rgfrosh: A Python package for calculating shock conditions using real gas equations of state
**Submitting author:** @corykinney (<a href="http://orcid.org/0000-0001-8774-3090">Cory Kinney</a>)
**Repository:** https://github.com/VasuLab/rgfrosh
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.1
