### 2024-06-03 #6806 @corykinney
I added a statement of need section to the paper. As for the missing DOIs, it seems that UCF's electronic dissertation website does not issue DOIs.
