### 2024-06-06 #6806 @kyleniemeyer
@blackspur thank you! We typically ask reviewers to complete their reviews within two or three weeks if possible, but this is flexible. In addition, you can perform the review over a period of time and check items off as you progress.

Our review guidelines are here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

In brief, you will work through a checklist of review items in the review issue on GitHub. We also encourage reviewers to open issues on the repository for the software, when needed, and leave comments in the review issue.
