### 2024-06-17 21abf52
Fix overwrite of parameterized test value
Closes #15
