### 2024-06-18 8a471b6
Remove erroneous test data
Comment out parametrized tests for ideal shock reflected temperature ratio corresponding to rows in Table II.3 of Gaydon and Hurle's The shock tube in high-temperature chemical physics with erroneous values that do not match hand calculations of the book's equations

Fixes #11
