### 2024-06-18 8628d04
Fix tests broken by NumPy 2.0
