### 2024-06-18 8fc373a
Remove reflected velocity ratio
Removed erroneous equation from Gaydon and Hurle's The shock tube in high-temperature chemical phyics for reflected shock velocity ratio

Fixes #5
