### 2024-06-21 e5bcbd7
Update paper to mention SDToolbox
Closes #14
