+++ paper.bib
@@ -18,6 +18,15 @@
   doi={10.1002/ijch.199600044}
 }
 
+@misc{sdtoolbox,
+  title={Shock and Detonation Toolbox},
+  author={Kao, S. and Ziegler, J. and Bitter, N. and Schmidt, B. and Lawson, J. and Shepherd, J. E.},
+  journal={California Institute of Technology, Pasadena, CA},
+  note = "February 11, 2023",
+  year=2023,
+  URL={https://shepherd.caltech.edu/EDL/PublicResources/sdt/}
+}
+
 @misc{cantera,
     author = "David G. Goodwin and Harry K. Moffat and Ingmar Schoegl and Raymond L.
               Speth and Bryan W. Weber",
