### 2024-06-21 89fc224
Update paper citation
Updated citation for SDToolbox based on README distributed with Python code
