+++ paper.bib
@@ -19,12 +19,12 @@
 }
 
 @misc{sdtoolbox,
-  title={Shock and Detonation Toolbox},
-  author={Kao, S. and Ziegler, J. and Bitter, N. and Schmidt, B. and Lawson, J. and Shepherd, J. E.},
-  journal={California Institute of Technology, Pasadena, CA},
-  note = "February 11, 2023",
-  year=2023,
-  URL={https://shepherd.caltech.edu/EDL/PublicResources/sdt/}
+  title = {SDToolbox - Numerical Tools for Shock and Detonation Wave Modeling},
+  author = {Browne, S. and Ziegler, J. and Bitter, N. and Schmidt, B. and Lawson, J. and Shepherd, J. E.},
+  publisher = {California Institute of Technology, Pasadena, CA},
+  note = "GALCIT Report FM2018.001",
+  year = 2021,
+  URL = {https://shepherd.caltech.edu/EDL/PublicResources/sdt/}
 }
 
 @misc{cantera,
