+++ mkdocs.yml
@@ -52,6 +52,7 @@ theme:
 
 plugins:
   - search
+  - social
   - mkdocstrings:
       handlers:
         python:
