### 2024-06-28 #6855 @cartemic
@kyleniemeyer It is going well, thank you for following up; my schedule has been hectic lately and this was a useful reminder that I forgot to leave a comment! I have reviewed the package and the paper and raised some issues, which @corykinney has addressed, and pending @blackspur's review I am happy to recommend `rgfrosh` for publication.
