### 2024-07-09 #6855 @blackspur
@kyleniemeyer apologies for the delay. Was caught up in some deadlines. I have finished my review and I am happy to recommend `rgfrosh` as well. The addition of shock property calculation for real gases to open-source is a welcome addition!
