+++ CHANGELOG.md
@@ -1,0 +1,7 @@
+## 0.2.2 
+2024-07-11
+
+### Removed
+
+- `IdealShock.reflected_velocity_ratio` function was removed due to an error in Gaydon and Hurle's derivation producing
+  incorrect results
