### 2024-07-11 ee0f47e
Update tests
Remove NumPy < 2.0 requirement
Add Python 3.12 support
Add Python 3.11 and 3.12 to tests to cover all officially supported versions
