+++ requirements.txt
@@ -1,4 +1,4 @@
-numpy < 2.0
+numpy
 cantera
 cairosvg
 CoolProp
