+++ codecov.yml
@@ -1,0 +1,2 @@
+codecov:
+ token: 14e3066d-4083-4080-843b-b28f0e31f804
