+++ rgfrosh/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.2.1"
+__version__ = "0.3.0"
 
 from .shock import IdealShock, FrozenShock
 from .errors import ConvergenceError
