+++ setup.cfg
@@ -2,7 +2,7 @@
 name = rgfrosh
 version = attr: rgfrosh.__version__
 author = Cory Kinney
-author_email = corykinney@ucf.edu
+author_email = cory.kinney@ucf.edu
 url = https://github.com/VasuLab/RGFROSH
 description = A frozen shock solver for ideal and real gas equations of state.
 long_description = file: README.md
