### 2024-07-11 7b0af48
Update CITATION.cff
