### 2024-07-11 #6855 @kyleniemeyer
Congratulations @corykinney on your article's publication in JOSS! (@corykinney I know you're already in the JOSS reviewer database, so I will probably be asking you to review something!)

Many thanks to @cartemic and @blackspur for reviewing this submission.
